chore: Export the table from the main `schema` file in hopes that fixes the issue
